refactor: move list item and filter out of SessionList
This is to keep the SessionList cleaner and separate the different
concerns.
